add verification before performing delete and edit sotre

+++ controllers/store.js
@@ -16,7 +16,8 @@ exports.store_create_post = (req, res) => {
 }
 
 // Edit store
-exports.store_edit_post = (req, res) => {
+exports.store_edit_post = async (req, res) => {
+    if(await hasPermission(req.user.id, req.query.id)) {
     console.log(`Editing store ${req.body.id}`);
     Store.findByIdAndUpdate(req.body._id, req.body, {new:true})
     .then((shop) => {
@@ -26,21 +27,30 @@ exports.store_edit_post = (req, res) => {
         console.log('An error occured while editing store');
         console.log(err);
         res.json({err});
-    })
+    })}else{
+        console.log(`user ${req.user.id} tried to edit a store ${req.query.id} which they dont own`);
+        res.json({"message": "You dont have permission to perform this action"});
+    }
 }
 
 // Delete store
-exports.store_delete_get = (req, res) => {
-    console.log(`Deleting store ${req.query.id}`);
-    Store.findByIdAndDelete(req.query.id)
-    .then((store) => {
-        res.json({store});
-    })
-    .catch((err) => {
-        console.log('An error occured while editing store');
-        console.log(err);
-        res.json({err});
-    })
+exports.store_delete_get = async (req, res) => {
+    if(await hasPermission(req.user.id, req.query.id)) {
+        console.log(`Deleting store ${req.query.id}`);
+        Store.findByIdAndDelete(req.query.id)
+        .then((store) => {
+            res.json({store});
+        })
+        .catch((err) => {
+            console.log('An error occured while editing store');
+            console.log(err);
+            res.json({err});
+        })
+    }
+    else{
+        console.log(`user ${req.user.id} tried to delete a store ${req.query.id}`);
+        res.json({"message": "You dont have permission to perform this action"});
+    }
 }
 
 // Get all stores
@@ -54,7 +64,6 @@ exports.store_getAll_get = (req, res) => {
         console.log(err);
         res.json({err});
     })
-
 }
 
 // Get one store
@@ -68,5 +77,19 @@ exports.store_detail_get = (req, res) => {
         console.log(err);
         res.json({err});
     })
-
 }
+
+//#region helper functions
+/**
+ * Check if a specific user owns the permissions to edit or delete a store
+ * @param {ID of the user} userId 
+ * @param {ID of the store} storeId 
+ */
+async function hasPermission(userId, storeId) {
+    const store = await Store.findById(storeId)
+        if(store){
+            return (store.user == userId);
+        }
+        return false;
+}
+//#endregion
